Function unitn_market_2022::market::market_test::test_sell_success

pub fn test_sell_success<T: Market>()

The sell function needs to be called after lock_sell, which creates an immutable contract ready to be fulfilled. This test assumes that lock_sell and get_sell_price work correctly, otherwise fails.

This test checks that:

the returned value is a positive or zero quantity of DEFAULT_GOOD_KIND.

Reference to the specs.

FSK team