Trait unitn_market_2022::market::Market

pub trait Market: Notifiable {
    fn new_random() -> Rc<RefCell<dyn Market>>
    where
        Self: Sized;
    fn new_with_quantities(
        eur: f32, 
        yen: f32, 
        usd: f32, 
        yuan: f32
    ) -> Rc<RefCell<dyn Market>>
    where
        Self: Sized;
    fn new_file(path: &str) -> Rc<RefCell<dyn Market>>
    where
        Self: Sized;
    fn get_name(&self) -> &'static str;
    fn get_budget(&self) -> f32;
    fn get_buy_price(
        &self, 
        kind: GoodKind, 
        quantity: f32
    ) -> Result<f32, MarketGetterError>;
    fn get_sell_price(
        &self, 
        kind: GoodKind, 
        quantity: f32
    ) -> Result<f32, MarketGetterError>;
    fn get_goods(&self) -> Vec<GoodLabel>;
    fn lock_buy(
        &mut self, 
        kind_to_buy: GoodKind, 
        quantity_to_buy: f32, 
        bid: f32, 
        trader_name: String
    ) -> Result<String, LockBuyError>;
    fn buy(&mut self, token: String, cash: &mut Good) -> Result<Good, BuyError>;
    fn lock_sell(
        &mut self, 
        kind_to_sell: GoodKind, 
        quantity_to_sell: f32, 
        offer: f32, 
        trader_name: String
    ) -> Result<String, LockSellError>;
    fn sell(
        &mut self, 
        token: String, 
        good: &mut Good
    ) -> Result<Good, SellError>;
}

This trait is to be implemented by each team

Required Methods

fn new_random() -> Rc<RefCell<dyn Market>> where
Self: Sized,

Initializes the market with random initial good amounts, whose total sum is crate::good::consts::STARTING_CAPITAL of the default good at the initial default rates

fn new_with_quantities(
eur: f32,
yen: f32,
usd: f32,
yuan: f32
) -> Rc<RefCell<dyn Market>> where
Self: Sized,

Initializes the market for debugging purposes, with fixed initial good amounts of 100000 of the default good at the initial default rates

fn new_file(path: &str) -> Rc<RefCell<dyn Market>> where
Self: Sized,

Initializes the market with initial good amounts as saved in the file at the provided path; if any i/o [Error] is encountered when reading the file, the function return a new instance created with new_random()

fn get_name(&self) -> &'static str

Documentation at: https://github.com/WG-AdvancedProgramming/market-protocol-specifications/blob/change_lock-buy_functions_logic/market-protocol-specifications.md#market-functionality

fn get_budget(&self) -> f32

fn get_buy_price(
&self,
kind: GoodKind,
quantity: f32
) -> Result<f32, MarketGetterError>

fn get_sell_price(
&self,
kind: GoodKind,
quantity: f32
) -> Result<f32, MarketGetterError>

fn get_goods(&self) -> Vec<GoodLabel>

fn lock_buy(
&mut self,
kind_to_buy: GoodKind,
quantity_to_buy: f32,
bid: f32,
trader_name: String
) -> Result<String, LockBuyError>

Documentation at: https://github.com/WG-AdvancedProgramming/market-protocol-specifications/blob/change_lock-buy_functions_logic/market-protocol-specifications.md#the-lock-function

fn buy(&mut self, token: String, cash: &mut Good) -> Result<Good, BuyError>

fn lock_sell(
&mut self,
kind_to_sell: GoodKind,
quantity_to_sell: f32,
offer: f32,
trader_name: String
) -> Result<String, LockSellError>

Documentation at: https://github.com/WG-AdvancedProgramming/market-protocol-specifications/blob/change_lock-buy_functions_logic/market-protocol-specifications.md#the-lock-function-1

fn sell(&mut self, token: String, good: &mut Good) -> Result<Good, SellError>

Documentation at: https://github.com/WG-AdvancedProgramming/market-protocol-specifications/blob/change_lock-buy_functions_logic/market-protocol-specifications.md#the-sell-function